requirements.txt
flask
line-bot-sdk
pytesseract
Pillow
gspread
oauth2client
